Resolve CI Rustfmt failure by formatting Rust sources

+++ src/helpers/utils.rs
@@ -96,7 +96,7 @@ pub fn print_device_detail(device: AllDevices) {
         &mut table,
         "Last Updated",
         device.last_update_time.map(|timestamp| {
-            let datetime = DateTime::from_timestamp_millis(timestamp as i64)
+            let datetime = DateTime::from_timestamp_millis(timestamp)
                 .unwrap_or_else(|| DateTime::from_timestamp(0, 0).expect("REASON"));
             datetime.format("%Y-%m-%d %H:%M:%S").to_string()
         }),
@@ -1152,7 +1152,9 @@ pub fn print_client_list(items: Vec<ClientListItem>) {
             item.connection_status.as_deref().unwrap_or("N/A"),
             health_score,
             ssid,
-            item.connected_network_device_name.as_deref().unwrap_or("N/A"),
+            item.connected_network_device_name
+                .as_deref()
+                .unwrap_or("N/A"),
         ]);
     }
     table.set_format(*format::consts::FORMAT_NO_BORDER_LINE_SEPARATOR);
